Refactor tiling setup
Setup for the tiling was happening in the viewDidLoad() function
inside the HistoricalViewController, so I moved that out so that
it could also be used by the quest view.
